redis & session removed & vitest setup
