+++ Dockerfile
@@ -8,7 +8,6 @@ COPY tsconfig.json .
 COPY vite.config.ts .
 
 RUN yarn install
-RUN mkdir node_modules/.vite/deps_temp && chmod -R 777 node_modules/.vite/deps_temp
 
 COPY . .
