+++ src/controller/PostController/PostController.test.ts
@@ -1,0 +1,6 @@
+import { describe, it, expect } from 'vitest'
+import request from 'supertest'
+
+describe('Testing Post Controller', () => {
+    it('Should return all post on / route', async () => {})
+})
