+++ src/controller/PostController/PostController.ts
@@ -11,7 +11,7 @@ export const getAllPosts = async (req: Request, res: Response) => {
 export const getPost = async (req: Request, res: Response) => {
     const id = req.params.id
     try {
-        const post = await PostModel.findById(id)
+        const post = await PostModel.findById(id).wtimeout(10)
         if (!post) {
             res.status(404).json({
                 type: 'error',
@@ -20,7 +20,7 @@ export const getPost = async (req: Request, res: Response) => {
             return
         }
         res.status(200).json({
-            type: 'success',
+            type: 'double success',
             data: post
         })
     } catch (err: unknown) {
