+++ src/utilities/index.ts
@@ -1,2 +1,1 @@
 export * from '../middlewares/rateLimiter'
-export * from './redis-utils'
